Added ability to cancel tool calls.
The framework for cancelling calls is ready. The shell tool is
currently the only one that has been modified to allow cancellation.

Removed the :stopped status and added the :cleanup status.
:cleanup is entered after normal tool completion or after interrupted
tool completion.  In either case, :cleanup removes the future and
resources from the state.  This is a failsafe in case the tool is not
able to clean up after itself.
This introduced another status, :cleanup, and will require
removing :stopped as a terminal status.

Since a deref of a cancelled future *immediately* results in a
CancellationException without waiting for the future to cleanup, we
needed to add another promise to the tool call state to ensure that we
properly wait for the tool call to cleanup on cancellation.  We added
that promise to the state, and the associated actions to init and
remove it from the state, and to deliver to the promise.  Added some
debug logging for the new promise.  Modified the tests to handle the
change in actions.

Added :future-cleanup-complete?* to the tool call state and associated
actions.

The :stopping status is now considered to be an active status.  Added
a self-transition on :stop-requested.

Added actions for cleaning up resource and futures.
Changed :set-future to :add-future.

Added a :cancel-future action to the :executing status on
:stop-requested.

Renamed the :set-call-future action to :set-future, for consistency.

In the shell tool, add and remove the resource (the spawned process)
from the call state.  On InterruptedException, destroy the process.

Added a tool-call-destroy-resource! multi-method that can be used from
inside and outside the tool to destroy any resources.  Typically, a
tool will clean up after itself.  But it might not, for example if it
never gets an InterruptedException, or if the tool is not cooperating.

Made the shell tool more robust in making sure that it always cleans
up and resources that remain at the end.  As a failsafe.

Marginally improved some logging.

Updated and added to the tests.