Improved flaky test.
The tool-calls-with-prompt-stop-test was falky in CI/CD.
Likely because it relied on sleep times.
Added promises to the test to coordinate between the concurrent tool
calls and the prompt stop.
Hopefully, this works better.

Changed tool-calls-with-prompt-stop to
tool-calls-with-prompt-stop-test for consistency.

Fixed a reflection warning in deep-sleep by casting the sleep arg to
long.